move helper functions/classes to rho_tau.py and plots.py
