+++ src/sp_validation/cosmo_val.py
@@ -1,18 +1,12 @@
 # %%
 import os
-import re
-import time
 from collections import Counter
-from contextlib import contextmanager
 
 import camb
 import colorama
 import healpy as hp
 import healsparse as hsp
 import matplotlib.pyplot as plt
-import matplotlib.scale as mscale
-import matplotlib.ticker as ticker
-import matplotlib.transforms as mtransforms
 import numpy as np
 import pymaster as nmt
 import skyproj
@@ -20,1094 +14,21 @@ import treecorr
 import yaml
 from astropy import units as u
 from astropy.coordinates import SkyCoord
+from astropy.cosmology import Planck18
 from astropy.io import fits
 from cosmo_numba.B_modes.schneider2022 import get_pure_EB_modes
 from cs_util import plots as cs_plots
 from shear_psf_leakage import leakage
 from shear_psf_leakage import plots as psfleak_plots
 from shear_psf_leakage import run_object, run_scale
-from shear_psf_leakage.rho_tau_cov import CovTauTh
-from shear_psf_leakage.rho_tau_stat import PSFErrorFit, RhoStat, TauStat
+from shear_psf_leakage.rho_tau_stat import PSFErrorFit
 from uncertainties import ufloat
 
-
-class SquareRootScale(mscale.ScaleBase):
-    """
-    ScaleBase class for generating square root scale.
-
-    Usage example: axis.set_yscale('squareroot')
-
-    """
-
-    name = "squareroot"
-
-    def __init__(self, axis, **kwargs):
-        mscale.ScaleBase.__init__(self, axis, **kwargs)
-
-    def set_default_locators_and_formatters(self, axis):
-        axis.set_major_locator(ticker.AutoLocator())
-        axis.set_major_formatter(ticker.ScalarFormatter())
-        axis.set_minor_locator(ticker.NullLocator())
-        axis.set_minor_formatter(ticker.NullFormatter())
-
-    def limit_range_for_scale(self, vmin, vmax, minpos):
-        return max(0.0, vmin), vmax
-
-    class SquareRootTransform(mtransforms.Transform):
-        input_dims = 1
-        output_dims = 1
-        is_separable = True
-
-        def transform_non_affine(self, a):
-            return np.array(a) ** 0.5
-
-        def inverted(self):
-            return SquareRootScale.InvertedSquareRootTransform()
-
-    class InvertedSquareRootTransform(mtransforms.Transform):
-        input_dims = 1
-        output_dims = 1
-        is_separable = True
-
-        def transform(self, a):
-            return np.array(a) ** 2
-
-        def inverted(self):
-            return SquareRootScale.SquareRootTransform()
-
-    def get_transform(self):
-        return self.SquareRootTransform()
-
-
-mscale.register_scale(SquareRootScale)
-
-not_square_size = [
-    "DES",
-    "SP_v1.3_LFmask_8k",
-    "SP_v1.3_LFmask_8k_no_alpha",
-    "SP_v1.3_LFmask_8k_li_2024",
-    "SP_v1.3_LFmask_8k_SN8",
-    "SP_v1.3_LFmask_8k_F2",
-]
-
-
-def get_params_rho_tau(cat, survey="other"):
-
-    # Set parameters
-    params = {}
-    # TODO to yaml file
-    if survey == "DES":
-        params["patch_number"] = 120
-        print("DES, jackknife patch number = 120")
-    elif survey == "SP_axel_v0.0":
-        params["patch_number"] = 120
-        print("SP_Axel_v0.0, jackknife patch number =120")
-    elif survey == "SP_v1.4-P3" or survey == "SP_v1.4-P3_LFmask":
-        params["patch_number"] = 120
-        print("SP_v1.4, jackknife patch number =120")
-    else:
-        params["patch_number"] = 150
-    params["ra_col"] = cat["psf"]["ra_col"]
-    params["dec_col"] = cat["psf"]["dec_col"]
-    params["e1_PSF_col"] = cat["psf"]["e1_PSF_col"]
-    params["e2_PSF_col"] = cat["psf"]["e2_PSF_col"]
-    params["e1_star_col"] = cat["psf"]["e1_star_col"]
-    params["e2_star_col"] = cat["psf"]["e2_star_col"]
-    params["PSF_size"] = cat["psf"]["PSF_size"]
-    params["star_size"] = cat["psf"]["star_size"]
-    params["square_size"] = survey not in not_square_size
-    if survey != "DES":
-        params["PSF_flag"] = cat["psf"]["PSF_flag"]
-        params["star_flag"] = cat["psf"]["star_flag"]
-    params["ra_units"] = "deg"
-    params["dec_units"] = "deg"
-
-    params["w_col"] = cat["shear"]["w"]
-    params["e1_col"] = cat["shear"]["e1_col"]
-    params["e2_col"] = cat["shear"]["e2_col"]
-    params["R11"] = cat["shear"].get("R11")
-    params["R22"] = cat["shear"].get("R22")
-
-    return params
-
-
-def get_rho_tau_w_cov(config, version, treecorr_config, outdir, method, cov_rho=False):
-    """
-    Method to compute the covariance matrices of rho and tau-statistics of a given list of versions in cosmo_val.
-    Also computes rho and tau-statistics.
-
-    Parameters
-    ----------
-    versions : list
-        List of versions to compute the covariance matrices for.
-    method : str
-        Method to compute the covariance matrices. Options are 'jk' or 'th'.
-    """
-    if method == "th":
-        nbin_ang, nbin_rad = 100, 200
-        rho_stat_handler, tau_stat_handler = get_rho_tau(
-            config, version, treecorr_config, outdir, cov_rho=cov_rho
-        )
-        get_theory_cov(
-            config,
-            version,
-            treecorr_config,
-            outdir,
-            nbin_ang=nbin_ang,
-            nbin_rad=nbin_rad,
-        )
-        return rho_stat_handler, tau_stat_handler
-    elif method == "jk":
-        return get_jackknife_cov(config, version, treecorr_config, outdir)
-    elif method == "sim":
-        if os.path.exists(outdir + "/cov_tau_" + version + "_th.npy"):
-            print(
-                f"Covariance from simulation available at the following file: {outdir+'/cov_tau_'+version+'_th.npy'}"
-            )
-            print(f"Computing rho and tau statistics for the version: {version}")
-            return get_rho_tau(
-                config, version, treecorr_config, outdir, cov_rho=cov_rho
-            )
-        else:
-            raise ValueError(
-                "Covariance from simulation not available. Please compute it first."
-            )
-    else:
-        raise ValueError("Method must be either 'jk' or 'th' or 'sim'.")
-
-
-def get_rho_tau(config, version, treecorr_config, outdir, cov_rho=False):
-    """
-    Compute rho and tau statistics for a given version of the catalogue.
-
-    Parameters
-    ----------
-    config: dict
-        Configuration file.
-    version : str
-        Version of the catalogue to use.
-    treecorr_config : dict
-        Configuraion for treecorr.
-    outdir : str
-        Output directory.
-    """
-
-    params = get_params_rho_tau(config[version], survey=version)
-
-    print("Compute Rho and Tau statistics for the version: ", version)
-    start_time = time.time()
-
-    out_base = f"rho_stats_{version}.fits"
-    out_path = f"{outdir}/{out_base}"
-
-    rho_stat_handler = RhoStat(
-        output=outdir, treecorr_config=treecorr_config, verbose=True
-    )
-
-    if os.path.exists(out_path):
-        print(f"Skipping rho statistics computation, file {out_path} already exists.")
-        rho_stat_handler.load_rho_stats(out_base)
-    else:
-
-        rho_stat_handler.catalogs.set_params(params, outdir)
-
-        mask = version != "DES"
-        square_size = params["square_size"]
-
-        # Build catalogues
-        rho_stat_handler.build_cat_to_compute_rho(
-            config[version]["psf"]["path"],
-            catalog_id=version,
-            square_size=square_size,
-            mask=mask,
-            hdu=(
-                config[version]["psf"]["hdu"]
-                if config[version]["psf"]["hdu"] is not None
-                else 1
-            ),
-        )
-
-        if cov_rho:
-            if not os.path.exists(outdir + "/cov_rho_" + version + ".npy"):
-                only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-                rho_stat_handler.compute_rho_stats(
-                    version,
-                    out_base,
-                    save_cov=True,
-                    func=only_p,
-                    var_method="jackknife",
-                )
-        else:
-            rho_stat_handler.compute_rho_stats(version, out_base, var_method=None)
-
-    out_base = f"tau_stats_{version}.fits"
-    out_path = f"{outdir}/{out_base}"
-
-    tau_stat_handler = TauStat(
-        catalogs=rho_stat_handler.catalogs,
-        output=outdir,
-        treecorr_config=treecorr_config,
-        verbose=True,
-    )
-
-    if os.path.exists(out_path):
-        print(f"Skipping tau statistics computation, file {out_path} already exists.")
-        tau_stat_handler.load_tau_stats(out_base)
-    else:
-
-        tau_stat_handler.catalogs.set_params(params, outdir)
-
-        mask = version != "DES"
-
-        square_size = params["square_size"]
-
-        # Build the different catalogs if necessary
-        if f"psf_{version}" not in tau_stat_handler.catalogs.catalogs_dict.keys():
-            tau_stat_handler.build_cat_to_compute_tau(
-                config[version]["psf"]["path"],
-                cat_type="psf",
-                catalog_id=version,
-                square_size=square_size,
-                mask=mask,
-                hdu=(
-                    config[version]["psf"]["hdu"]
-                    if config[version]["psf"]["hdu"] is not None
-                    else 1
-                ),
-            )
-
-        # Build the catalog of galaxies. PSF was computed above
-        tau_stat_handler.build_cat_to_compute_tau(
-            config[version]["shear"]["path"],
-            cat_type="gal",
-            catalog_id=version,
-            square_size=square_size,
-            mask=mask,
-        )
-
-        # function to extract the tau_+
-        tau_stat_handler.compute_tau_stats(version, out_base, var_method=None)
-
-    print(f"Time to compute rho and tau statistics: {time.time() - start_time:.2f} s")
-    return rho_stat_handler, tau_stat_handler
-
-
-def get_theory_cov(
-    config, version, treecorr_config, outdir, nbin_ang=100, nbin_rad=100
-):
-    """
-    Compute an analytical estimate of the covariance matrix of rho and tau-statistics.
-    """
-
-    params = get_params_rho_tau(config[version], survey=version)
-
-    info = config[version]
-    A = info["cov_th"]["A"] * 60 * 60
-    n_e = info["cov_th"]["n_e"]
-    n_psf = info["cov_th"]["n_psf"]
-
-    path_gal = info["shear"]["path"]
-    path_psf = info["psf"]["path"]
-    hdu_psf = info["psf"]["hdu"]
-
-    print("Computing the covariance matrix for the version: ", version)
-    start_time = time.time()
-
-    if os.path.exists(outdir + "/cov_tau_" + version + "_th.npy"):
-        print(
-            f"Skipping covariance computation, file {outdir+'/cov_tau_'+version+'_th.npy'} already exists."
-        )
-        return
-
-    cov_tau_th = CovTauTh(
-        path_gal=path_gal,
-        path_psf=path_psf,
-        hdu_psf=hdu_psf,
-        treecorr_config=treecorr_config,
-        A=A,
-        n_e=n_e,
-        n_psf=n_psf,
-        params=params,
-    )
-
-    print(
-        "--- Computation of the rho and tau statistics for the covariance %s seconds ---"
-        % (time.time() - start_time)
-    )
-
-    cov = cov_tau_th.build_cov(nbin_ang=nbin_ang, nbin_rad=nbin_rad)
-    print("--- Covariance computation %s seconds ---" % (time.time() - start_time))
-    np.save(outdir + "/cov_tau_" + version + "_th.npy", cov)
-    print("Saved covariance matrix of version: ", version)
-    del cov_tau_th
-    return
-
-
-def get_jackknife_cov(config, version, treecorr_config, outdir, ncov=100):
-    """
-    Compute the covariance matrix of rho and tau-statistics using the jackknife method.
-    Also compute rho and tau-statistics.
-    """
-
-    if os.path.exists(outdir + "/cov_tau_" + version + "_jk.npy"):
-        print(
-            f"Skipping covariance computation, file {outdir+'/cov_tau_'+version+'_jk.npy'} already exists."
-        )
-        rho_stat_handler = RhoStat(
-            output=outdir, treecorr_config=treecorr_config, verbose=False
-        )
-
-        tau_stat_handler = TauStat(
-            catalogs=rho_stat_handler.catalogs,
-            output=outdir,
-            treecorr_config=treecorr_config,
-            verbose=True,
-        )
-
-        return rho_stat_handler, tau_stat_handler
-
-    for i in range(ncov):
-
-        if not (
-            os.path.exists(outdir + "/cov_tau_" + version + str(i) + ".npy")
-            and os.path.exists(outdir + "/cov_rho_" + version + str(i) + ".npy")
-        ):
-
-            params = get_params_rho_tau(config[version], survey=version)
-
-            rho_stat_handler = RhoStat(
-                output=outdir, treecorr_config=treecorr_config, verbose=False
-            )
-
-            out_base = f"rho_stats_{version}.fits"
-            out_path = f"{outdir}/{out_base}"
-
-            print(
-                f"Computing rho-statistics of version {version} for jackknife patch {i+1}/{ncov}"
-            )
-
-            rho_stat_handler.catalogs.set_params(params, outdir)
-
-            mask = version != "DES"
-            square_size = params["square_size"]
-
-            # Build catalogues
-            rho_stat_handler.build_cat_to_compute_rho(
-                config[version]["psf"]["path"],
-                catalog_id=version + str(i),
-                square_size=square_size,
-                mask=mask,
-                hdu=config[version]["psf"]["hdu"],
-            )
-
-            # Compute and save rho stats
-            only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-            rho_stat_handler.compute_rho_stats(
-                version + str(i),
-                out_base,
-                save_cov=True,
-                func=only_p,
-                var_method="jackknife",
-            )
-
-            tau_stat_handler = TauStat(
-                catalogs=rho_stat_handler.catalogs,
-                output=outdir,
-                treecorr_config=treecorr_config,
-                verbose=True,
-            )
-
-            out_base = f"tau_stats_{version}.fits"
-            out_path = f"{outdir}/{out_base}"
-
-            tau_stat_handler.catalogs.set_params(params, outdir)
-
-            mask = version != "DES"
-
-            square_size = params["square_size"]
-
-            # Build the different catalogs if necessary
-            if (
-                f"psf_{version}{i}"
-                not in tau_stat_handler.catalogs.catalogs_dict.keys()
-            ):
-                tau_stat_handler.build_cat_to_compute_tau(
-                    config[version]["psf"]["path"],
-                    cat_type="psf",
-                    catalog_id=version,
-                    square_size=square_size,
-                    mask=mask,
-                )
-
-            # Build the catalog of galaxies. PSF was computed above
-            tau_stat_handler.build_cat_to_compute_tau(
-                config[version]["shear"]["path"],
-                cat_type="gal",
-                catalog_id=version + str(i),
-                square_size=square_size,
-                mask=mask,
-            )
-
-            # function to extract the tau_+
-            only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-            tau_stat_handler.compute_tau_stats(
-                version + str(i),
-                out_base,
-                save_cov=True,
-                func=only_p,
-                var_method="jackknife",
-            )
-
-            if (i + 1) != ncov:
-                del rho_stat_handler, tau_stat_handler
-
-    cov_tau_loc = np.zeros((60, 60))
-    cov_rho_loc = np.zeros((120, 120))
-    for i in range(ncov):
-        cov_tau_loc += np.load(outdir + f"/cov_tau_{version}{i}.npy")
-        cov_rho_loc += np.load(outdir + f"/cov_rho_{version}{i}.npy")
-        os.remove(outdir + f"/cov_tau_{version}{i}.npy")
-        os.remove(outdir + f"/cov_rho_{version}{i}.npy")
-
-    cov_tau = cov_tau_loc / ncov
-    cov_rho = cov_rho_loc / ncov
-
-    np.save(outdir + "/cov_tau_" + version + "_jk.npy", cov_tau)
-    np.save(outdir + "/cov_rho_" + version + ".npy", cov_rho)
-
-    return rho_stat_handler, tau_stat_handler
-
-
-def get_samples(psf_fitter, version, cov_type="jk", apply_debias=None, sampler="emcee"):
-    """
-    Samples (alpha, beta, eta) using the sampler 'emcee' or 'lsq'
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    cov_type : str
-        Type of covariance matrix to use. Options are 'jk' or 'th' or 'sim'.
-    apply_debias : int
-        If not None, apply debiasing to the least square method.
-    sampler : str
-        Sampler to use. Options are 'emcee' or 'lsq'. (Default: 'emcee')
-    """
-    if sampler == "emcee":
-        return get_samples_emcee(
-            psf_fitter, version, cov_type=cov_type, apply_debias=apply_debias
-        )
-    elif sampler == "lsq":
-        return get_samples_lsq(
-            psf_fitter, version, cov_type=cov_type, apply_debias=apply_debias
-        )
-    else:
-        raise ValueError("Sampler must be either 'emcee' or 'lsq'.")
-
-
-def get_samples_emcee(
-    psf_fitter, version, nwalkers=124, nsamples=10000, cov_type="jk", apply_debias=None
-):
-    """
-    Samples (alpha, beta, eta) using the covariance of the tau statistics and emcee.
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    nwalkers : int
-        Number of walkers to use in the MCMC. (Default: 124)
-    nsamples : int
-        Number of samples to draw from the MCMC. (Default 10000)
-    cov_type : str
-        Type of covariance matrix to use. Options are 'jk' or 'th' or 'sim'. (Default: 'jk')
-    """
-    # Load rho and tau stats
-    psf_fitter.load_rho_stat("rho_stats_" + version + ".fits")
-    psf_fitter.load_tau_stat("tau_stats_" + version + ".fits")
-
-    # Check if the path exists
-    sample_file_path = psf_fitter.get_sample_file_path(version)
-    if os.path.exists(sample_file_path):
-        print(f"Skipping sample computation, file {sample_file_path} already exists.")
-        flat_samples = psf_fitter.load_samples(version)
-        mcmc_result, q = psf_fitter.get_mcmc_from_samples(version)
-        print(mcmc_result)
-    # Or run MCMC
-    else:
-        print("MCMC sampling")
-        psf_fitter.load_covariance("cov_tau_" + version + "_" + cov_type + ".npy")
-
-        npatch = apply_debias if (apply_debias is not None) else None
-        flat_samples, mcmc_result, q = psf_fitter.run_chain(
-            nwalkers=nwalkers,
-            nsamples=nsamples,
-            npatch=npatch,
-            apply_debias=npatch is not None,
-            savefig="mcmc_samples_" + version + ".png",
-        )
-        psf_fitter.save_samples(flat_samples, version)
-    return flat_samples, mcmc_result, q
-
-
-def get_samples_lsq(psf_fitter, version, apply_debias=None, cov_type="jk"):
-    """
-    Samples (alpha, beta, eta) using the covariance of the tau statistics and least square method.
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    apply_debias : int
-        If not None, apply debiasing to the least square method. (Default: None)
-    """
-    # Load rho and tau stats
-    psf_fitter.load_rho_stat("rho_stats_" + version + ".fits")
-    psf_fitter.load_tau_stat("tau_stats_" + version + ".fits")
-
-    # Check if the path exists
-    sample_file_path = psf_fitter.get_sample_path(version)
-    if os.path.exists(sample_file_path):
-        print(f"Skipping sample computation, file {sample_file_path} already exists.")
-        flat_samples = psf_fitter.load_samples(version)
-        mcmc_result, q = psf_fitter.get_mcmc_from_samples(flat_samples)
-        print(mcmc_result)
-    # Or run MCMC
-    else:
-        print("Least square sampling")
-        psf_fitter.load_covariance(
-            "cov_tau_" + version + "_" + cov_type + ".npy", cov_type="tau"
-        )
-        psf_fitter.load_covariance("cov_rho_" + version + ".npy", cov_type="rho")
-        npatch = apply_debias if (apply_debias is not None) else None
-        flat_samples, mcmc_result, q = psf_fitter.get_least_squares_params_samples(
-            npatch=npatch, apply_debias=(npatch is not None)
-        )
-        psf_fitter.save_samples(flat_samples, version)
-    return flat_samples, mcmc_result, q
-
-
-not_square_size = [
-    "DES",
-    "SP_v1.3_LFmask_8k",
-    "SP_v1.3_LFmask_8k_no_alpha",
-    "SP_v1.3_LFmask_8k_li_2024",
-    "SP_v1.3_LFmask_8k_SN8",
-    "SP_v1.3_LFmask_8k_F2",
-]
-
-
-def get_params_rho_tau(cat, survey="other"):
-
-    # Set parameters
-    params = {}
-    # TODO to yaml file
-    if survey == "DES":
-        params["patch_number"] = 120
-        print("DES, jackknife patch number = 120")
-    elif survey == "SP_axel_v0.0":
-        params["patch_number"] = 120
-        print("SP_Axel_v0.0, jackknife patch number =120")
-    elif survey == "SP_v1.4-P3" or survey == "SP_v1.4-P3_LFmask":
-        params["patch_number"] = 120
-        print("SP_v1.4, jackknife patch number =120")
-    else:
-        params["patch_number"] = 150
-    params["ra_col"] = cat["psf"]["ra_col"]
-    params["dec_col"] = cat["psf"]["dec_col"]
-    params["e1_PSF_col"] = cat["psf"]["e1_PSF_col"]
-    params["e2_PSF_col"] = cat["psf"]["e2_PSF_col"]
-    params["e1_star_col"] = cat["psf"]["e1_star_col"]
-    params["e2_star_col"] = cat["psf"]["e2_star_col"]
-    params["PSF_size"] = cat["psf"]["PSF_size"]
-    params["star_size"] = cat["psf"]["star_size"]
-    params["square_size"] = survey not in not_square_size
-    if survey != "DES":
-        params["PSF_flag"] = cat["psf"]["PSF_flag"]
-        params["star_flag"] = cat["psf"]["star_flag"]
-    params["ra_units"] = "deg"
-    params["dec_units"] = "deg"
-
-    params["w_col"] = cat["shear"]["w"]
-    params["e1_col"] = cat["shear"]["e1_col"]
-    params["e2_col"] = cat["shear"]["e2_col"]
-    params["R11"] = cat["shear"].get("R11")
-    params["R22"] = cat["shear"].get("R22")
-
-    return params
-
-
-def get_rho_tau_w_cov(config, version, treecorr_config, outdir, method, cov_rho=False):
-    """
-    Method to compute the covariance matrices of rho and tau-statistics of a given list of versions in cosmo_val.
-    Also computes rho and tau-statistics.
-
-    Parameters
-    ----------
-    versions : list
-        List of versions to compute the covariance matrices for.
-    method : str
-        Method to compute the covariance matrices. Options are 'jk' or 'th'.
-    """
-    if method == "th":
-        nbin_ang, nbin_rad = 100, 200
-        rho_stat_handler, tau_stat_handler = get_rho_tau(
-            config, version, treecorr_config, outdir, cov_rho=cov_rho
-        )
-        get_theory_cov(
-            config,
-            version,
-            treecorr_config,
-            outdir,
-            nbin_ang=nbin_ang,
-            nbin_rad=nbin_rad,
-        )
-        return rho_stat_handler, tau_stat_handler
-    elif method == "jk":
-        return get_jackknife_cov(config, version, treecorr_config, outdir)
-    elif method == "sim":
-        if os.path.exists(outdir + "/cov_tau_" + version + "_th.npy"):
-            print(
-                f"Covariance from simulation available at the following file: {outdir+'/cov_tau_'+version+'_th.npy'}"
-            )
-            print(f"Computing rho and tau statistics for the version: {version}")
-            return get_rho_tau(
-                config, version, treecorr_config, outdir, cov_rho=cov_rho
-            )
-        else:
-            raise ValueError(
-                "Covariance from simulation not available. Please compute it first."
-            )
-    else:
-        raise ValueError("Method must be either 'jk' or 'th' or 'sim'.")
-
-
-def get_rho_tau(config, version, treecorr_config, outdir, cov_rho=False):
-    """
-    Compute rho and tau statistics for a given version of the catalogue.
-
-    Parameters
-    ----------
-    config: dict
-        Configuration file.
-    version : str
-        Version of the catalogue to use.
-    treecorr_config : dict
-        Configuraion for treecorr.
-    outdir : str
-        Output directory.
-    """
-
-    params = get_params_rho_tau(config[version], survey=version)
-
-    print("Compute Rho and Tau statistics for the version: ", version)
-    start_time = time.time()
-
-    out_base = f"rho_stats_{version}.fits"
-    out_path = f"{outdir}/{out_base}"
-
-    rho_stat_handler = RhoStat(
-        output=outdir, treecorr_config=treecorr_config, verbose=True
-    )
-
-    if os.path.exists(out_path):
-        print(f"Skipping rho statistics computation, file {out_path} already exists.")
-        rho_stat_handler.load_rho_stats(out_base)
-    else:
-
-        rho_stat_handler.catalogs.set_params(params, outdir)
-
-        mask = version != "DES"
-        square_size = params["square_size"]
-
-        # Build catalogues
-        rho_stat_handler.build_cat_to_compute_rho(
-            config[version]["psf"]["path"],
-            catalog_id=version,
-            square_size=square_size,
-            mask=mask,
-            hdu=(
-                config[version]["psf"]["hdu"]
-                if config[version]["psf"]["hdu"] is not None
-                else 1
-            ),
-        )
-
-        if cov_rho:
-            if not os.path.exists(outdir + "/cov_rho_" + version + ".npy"):
-                only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-                rho_stat_handler.compute_rho_stats(
-                    version,
-                    out_base,
-                    save_cov=True,
-                    func=only_p,
-                    var_method="jackknife",
-                )
-        else:
-            rho_stat_handler.compute_rho_stats(version, out_base, var_method=None)
-
-    out_base = f"tau_stats_{version}.fits"
-    out_path = f"{outdir}/{out_base}"
-
-    tau_stat_handler = TauStat(
-        catalogs=rho_stat_handler.catalogs,
-        output=outdir,
-        treecorr_config=treecorr_config,
-        verbose=True,
-    )
-
-    if os.path.exists(out_path):
-        print(f"Skipping tau statistics computation, file {out_path} already exists.")
-        tau_stat_handler.load_tau_stats(out_base)
-    else:
-
-        tau_stat_handler.catalogs.set_params(params, outdir)
-
-        mask = version != "DES"
-
-        square_size = params["square_size"]
-
-        # Build the different catalogs if necessary
-        if f"psf_{version}" not in tau_stat_handler.catalogs.catalogs_dict.keys():
-            tau_stat_handler.build_cat_to_compute_tau(
-                config[version]["psf"]["path"],
-                cat_type="psf",
-                catalog_id=version,
-                square_size=square_size,
-                mask=mask,
-                hdu=(
-                    config[version]["psf"]["hdu"]
-                    if config[version]["psf"]["hdu"] is not None
-                    else 1
-                ),
-            )
-
-        # Build the catalog of galaxies. PSF was computed above
-        tau_stat_handler.build_cat_to_compute_tau(
-            config[version]["shear"]["path"],
-            cat_type="gal",
-            catalog_id=version,
-            square_size=square_size,
-            mask=mask,
-        )
-
-        # function to extract the tau_+
-        tau_stat_handler.compute_tau_stats(version, out_base, var_method=None)
-
-    print(f"Time to compute rho and tau statistics: {time.time() - start_time:.2f} s")
-    return rho_stat_handler, tau_stat_handler
-
-
-def get_theory_cov(
-    config, version, treecorr_config, outdir, nbin_ang=100, nbin_rad=100
-):
-    """
-    Compute an analytical estimate of the covariance matrix of rho and tau-statistics.
-    """
-
-    params = get_params_rho_tau(config[version], survey=version)
-
-    info = config[version]
-    A = info["cov_th"]["A"] * 60 * 60
-    n_e = info["cov_th"]["n_e"]
-    n_psf = info["cov_th"]["n_psf"]
-
-    path_gal = info["shear"]["path"]
-    path_psf = info["psf"]["path"]
-    hdu_psf = info["psf"]["hdu"]
-
-    print("Computing the covariance matrix for the version: ", version)
-    start_time = time.time()
-
-    if os.path.exists(outdir + "/cov_tau_" + version + "_th.npy"):
-        print(
-            f"Skipping covariance computation, file {outdir+'/cov_tau_'+version+'_th.npy'} already exists."
-        )
-        return
-
-    cov_tau_th = CovTauTh(
-        path_gal=path_gal,
-        path_psf=path_psf,
-        hdu_psf=hdu_psf,
-        treecorr_config=treecorr_config,
-        A=A,
-        n_e=n_e,
-        n_psf=n_psf,
-        params=params,
-    )
-
-    print(
-        "--- Computation of the rho and tau statistics for the covariance %s seconds ---"
-        % (time.time() - start_time)
-    )
-
-    cov = cov_tau_th.build_cov(nbin_ang=nbin_ang, nbin_rad=nbin_rad)
-    print("--- Covariance computation %s seconds ---" % (time.time() - start_time))
-    np.save(outdir + "/cov_tau_" + version + "_th.npy", cov)
-    print("Saved covariance matrix of version: ", version)
-    del cov_tau_th
-    return
-
-
-def get_jackknife_cov(config, version, treecorr_config, outdir, ncov=100):
-    """
-    Compute the covariance matrix of rho and tau-statistics using the jackknife method.
-    Also compute rho and tau-statistics.
-    """
-
-    if os.path.exists(outdir + "/cov_tau_" + version + "_jk.npy"):
-        print(
-            f"Skipping covariance computation, file {outdir+'/cov_tau_'+version+'_jk.npy'} already exists."
-        )
-        rho_stat_handler = RhoStat(
-            output=outdir, treecorr_config=treecorr_config, verbose=False
-        )
-
-        tau_stat_handler = TauStat(
-            catalogs=rho_stat_handler.catalogs,
-            output=outdir,
-            treecorr_config=treecorr_config,
-            verbose=True,
-        )
-
-        return rho_stat_handler, tau_stat_handler
-
-    for i in range(ncov):
-
-        if not (
-            os.path.exists(outdir + "/cov_tau_" + version + str(i) + ".npy")
-            and os.path.exists(outdir + "/cov_rho_" + version + str(i) + ".npy")
-        ):
-
-            params = get_params_rho_tau(config[version], survey=version)
-
-            rho_stat_handler = RhoStat(
-                output=outdir, treecorr_config=treecorr_config, verbose=False
-            )
-
-            out_base = f"rho_stats_{version}.fits"
-            out_path = f"{outdir}/{out_base}"
-
-            print(
-                f"Computing rho-statistics of version {version} for jackknife patch {i+1}/{ncov}"
-            )
-
-            rho_stat_handler.catalogs.set_params(params, outdir)
-
-            mask = version != "DES"
-            square_size = params["square_size"]
-
-            # Build catalogues
-            rho_stat_handler.build_cat_to_compute_rho(
-                config[version]["psf"]["path"],
-                catalog_id=version + str(i),
-                square_size=square_size,
-                mask=mask,
-                hdu=config[version]["psf"]["hdu"],
-            )
-
-            # Compute and save rho stats
-            only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-            rho_stat_handler.compute_rho_stats(
-                version + str(i),
-                out_base,
-                save_cov=True,
-                func=only_p,
-                var_method="jackknife",
-            )
-
-            tau_stat_handler = TauStat(
-                catalogs=rho_stat_handler.catalogs,
-                output=outdir,
-                treecorr_config=treecorr_config,
-                verbose=True,
-            )
-
-            out_base = f"tau_stats_{version}.fits"
-            out_path = f"{outdir}/{out_base}"
-
-            tau_stat_handler.catalogs.set_params(params, outdir)
-
-            mask = version != "DES"
-
-            square_size = params["square_size"]
-
-            # Build the different catalogs if necessary
-            if (
-                f"psf_{version}{i}"
-                not in tau_stat_handler.catalogs.catalogs_dict.keys()
-            ):
-                tau_stat_handler.build_cat_to_compute_tau(
-                    config[version]["psf"]["path"],
-                    cat_type="psf",
-                    catalog_id=version,
-                    square_size=square_size,
-                    mask=mask,
-                )
-
-            # Build the catalog of galaxies. PSF was computed above
-            tau_stat_handler.build_cat_to_compute_tau(
-                config[version]["shear"]["path"],
-                cat_type="gal",
-                catalog_id=version + str(i),
-                square_size=square_size,
-                mask=mask,
-            )
-
-            # function to extract the tau_+
-            only_p = lambda corrs: np.array([corr.xip for corr in corrs]).flatten()
-            tau_stat_handler.compute_tau_stats(
-                version + str(i),
-                out_base,
-                save_cov=True,
-                func=only_p,
-                var_method="jackknife",
-            )
-
-            if (i + 1) != ncov:
-                del rho_stat_handler, tau_stat_handler
-
-    cov_tau_loc = np.zeros((60, 60))
-    cov_rho_loc = np.zeros((120, 120))
-    for i in range(ncov):
-        cov_tau_loc += np.load(outdir + f"/cov_tau_{version}{i}.npy")
-        cov_rho_loc += np.load(outdir + f"/cov_rho_{version}{i}.npy")
-        os.remove(outdir + f"/cov_tau_{version}{i}.npy")
-        os.remove(outdir + f"/cov_rho_{version}{i}.npy")
-
-    cov_tau = cov_tau_loc / ncov
-    cov_rho = cov_rho_loc / ncov
-
-    np.save(outdir + "/cov_tau_" + version + "_jk.npy", cov_tau)
-    np.save(outdir + "/cov_rho_" + version + ".npy", cov_rho)
-
-    return rho_stat_handler, tau_stat_handler
-
-
-def get_samples(psf_fitter, version, cov_type="jk", apply_debias=None, sampler="emcee"):
-    """
-    Samples (alpha, beta, eta) using the sampler 'emcee' or 'lsq'
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    cov_type : str
-        Type of covariance matrix to use. Options are 'jk' or 'th' or 'sim'.
-    apply_debias : int
-        If not None, apply debiasing to the least square method.
-    sampler : str
-        Sampler to use. Options are 'emcee' or 'lsq'. (Default: 'emcee')
-    """
-    if sampler == "emcee":
-        return get_samples_emcee(
-            psf_fitter, version, cov_type=cov_type, apply_debias=apply_debias
-        )
-    elif sampler == "lsq":
-        return get_samples_lsq(
-            psf_fitter, version, cov_type=cov_type, apply_debias=apply_debias
-        )
-    else:
-        raise ValueError("Sampler must be either 'emcee' or 'lsq'.")
-
-
-def get_samples_emcee(
-    psf_fitter, version, nwalkers=124, nsamples=10000, cov_type="jk", apply_debias=None
-):
-    """
-    Samples (alpha, beta, eta) using the covariance of the tau statistics and emcee.
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    nwalkers : int
-        Number of walkers to use in the MCMC. (Default: 124)
-    nsamples : int
-        Number of samples to draw from the MCMC. (Default 10000)
-    cov_type : str
-        Type of covariance matrix to use. Options are 'jk' or 'th' or 'sim'. (Default: 'jk')
-    """
-    # Load rho and tau stats
-    psf_fitter.load_rho_stat("rho_stats_" + version + ".fits")
-    psf_fitter.load_tau_stat("tau_stats_" + version + ".fits")
-
-    # Check if the path exists
-    sample_file_path = psf_fitter.get_sample_file_path(version)
-    if os.path.exists(sample_file_path):
-        print(f"Skipping sample computation, file {sample_file_path} already exists.")
-        flat_samples = psf_fitter.load_samples(version)
-        mcmc_result, q = psf_fitter.get_mcmc_from_samples(version)
-        print(mcmc_result)
-    # Or run MCMC
-    else:
-        print("MCMC sampling")
-        psf_fitter.load_covariance("cov_tau_" + version + "_" + cov_type + ".npy")
-
-        npatch = apply_debias if (apply_debias is not None) else None
-        flat_samples, mcmc_result, q = psf_fitter.run_chain(
-            nwalkers=nwalkers,
-            nsamples=nsamples,
-            npatch=npatch,
-            apply_debias=npatch is not None,
-            savefig="mcmc_samples_" + version + ".png",
-        )
-        psf_fitter.save_samples(flat_samples, version)
-    return flat_samples, mcmc_result, q
-
-
-def get_samples_lsq(psf_fitter, version, apply_debias=None, cov_type="jk"):
-    """
-    Samples (alpha, beta, eta) using the covariance of the tau statistics and least square method.
-
-    Parameters
-    ----------
-    psf_fitter : PSFFitter
-        Instance of PSFFitter.
-    version : str
-        Version of the catalogue to use.
-    apply_debias : int
-        If not None, apply debiasing to the least square method. (Default: None)
-    """
-    # Load rho and tau stats
-    psf_fitter.load_rho_stat("rho_stats_" + version + ".fits")
-    psf_fitter.load_tau_stat("tau_stats_" + version + ".fits")
-
-    # Check if the path exists
-    sample_file_path = psf_fitter.get_sample_path(version)
-    if os.path.exists(sample_file_path):
-        print(f"Skipping sample computation, file {sample_file_path} already exists.")
-        flat_samples = psf_fitter.load_samples(version)
-        mcmc_result, q = psf_fitter.get_mcmc_from_samples(flat_samples)
-        print(mcmc_result)
-    # Or run MCMC
-    else:
-        print("Least square sampling")
-        psf_fitter.load_covariance(
-            "cov_tau_" + version + "_" + cov_type + ".npy", cov_type="tau"
-        )
-        psf_fitter.load_covariance("cov_rho_" + version + ".npy", cov_type="rho")
-        npatch = apply_debias if (apply_debias is not None) else None
-        flat_samples, mcmc_result, q = psf_fitter.get_least_squares_params_samples(
-            npatch=npatch, apply_debias=(npatch is not None)
-        )
-        psf_fitter.save_samples(flat_samples, version)
-    return flat_samples, mcmc_result, q
+from .rho_tau import get_params_rho_tau, get_rho_tau_w_cov, get_samples
 
 
 # %%
 class CosmologyValidation:
-
     def __init__(
         self,
         versions,
@@ -1407,7 +328,7 @@ class CosmologyValidation:
         params["ra_units"] = "deg"
         params["dec_units"] = "deg"
 
-        params["w_col"] = self.cc[ver]["shear"]["w_col"]
+        params["w_col"] = self.cc[survey]["shear"]["w_col"]
 
         return params
 
@@ -3122,275 +2043,3 @@ class CosmologyValidation:
         plt.suptitle("Pseudo-Cl BB (Gaussian covariance)")
         plt.legend()
         plt.savefig(out_path)
-
-
-class FootprintPlotter:
-    """Class to create footprint plots.
-
-    Parameters
-    -----------
-    nside_coverage: int, optional
-        basic resolution of map; default is 32
-    nside_map:
-        fine resolution for plotting; default is 2048
-
-    """
-
-    # Dictionary storing region parameters
-    _regions = {
-        "NGC": {"ra_0": 180, "extend": [120, 270, 20, 70], "vmax": 60},
-        "SGC": {"ra_0": 15, "extend": [-20, 45, 20, 45], "vmax": 60},
-        "fullsky": {"ra_0": 150, "extend": [0, 360, -90, 90], "vmax": 60},
-    }
-
-    def __init__(self, nside_coverage=32, nside_map=2048):
-
-        self._nside_coverage = nside_coverage
-        self._nside_map = nside_map
-
-    def create_hsp_map(self, ra, dec):
-        """Create Hsp Map.
-
-        Create healsparse map.
-
-        Parameters
-        ----------
-        ra : numpy.ndarray
-            right ascension values
-        dec : numpy.ndarray
-            declination values
-
-        Returns
-        -------
-        hsp.HealSparseMap
-            map
-
-        """
-        # Create empty map
-        hsp_map = hsp.HealSparseMap.make_empty(
-            self._nside_coverage, self._nside_map, dtype=np.float32, sentinel=np.nan
-        )
-
-        # Get pixel list corresponding to coordinates
-        hpix = hp.ang2pix(self._nside_map, ra, dec, nest=True, lonlat=True)
-
-        # Get count of objects per pixel
-        pixel_counts = Counter(hpix)
-
-        # List of unique pixels
-        unique_hpix = np.array(list(pixel_counts.keys()))
-
-        # Number of objects
-        values = np.array(list(pixel_counts.values()), dtype=np.float32)
-
-        # Create maps with numbers per pixel
-        hsp_map[unique_hpix] = values
-
-        return hsp_map
-
-    def plot_area(
-        self,
-        hsp_map,
-        ra_0=0,
-        extend=[120, 270, 29, 70],
-        vmax=60,
-        projection=None,
-        outpath=None,
-        title=None,
-    ):
-        """Plot Area.
-
-        Plot catalogue in an area on the sky.
-
-        Parameters
-        ----------
-        hsp_map : hsp_HealSparseMap
-            input map
-        ra_0 : float, optional
-            anchor point in R.A.; default is 0
-        extend : list, optional
-            sky region, extend=[ra_low, ra_high, dec_low, dec_high];
-            default is [120, 270, 29, 70]
-        vmax : float, optional
-            maximum pixel value to plot with color; default is 60
-        projection : skyproj.McBrydeSkyproj
-            if ``None`` (default), a new plot is created
-        outpath : str, optional
-            output path, default is ``None``
-        title : str, optional
-            print title if not ``None`` (default)
-
-        Returns
-        --------
-        skyproj.McBrydeSkyproj
-            projection instance
-        plt.axes.Axes
-            axes instance
-
-        Raises
-        ------
-        ValueError
-            if no object found in region
-
-        """
-        if not projection:
-
-            # Create new figure and axes
-            fig, ax = plt.subplots(figsize=(10, 10))
-
-            # Create new projection
-            projection = skyproj.McBrydeSkyproj(
-                ax=ax, lon_0=ra_0, extent=extend, autorescale=True, vmax=vmax
-            )
-        else:
-            ax = None
-
-        try:
-            _ = projection.draw_hspmap(
-                hsp_map, lon_range=extend[0:2], lat_range=extend[2:]
-            )
-        except ValueError:
-            msg = "No object found in region to draw"
-            print(f"{msg}, continuing...")
-            # raise ValueError(msg)
-
-        projection.draw_milky_way(width=25, linewidth=1.5, color="black", linestyle="-")
-
-        if title:
-            plt.title(title, pad=5)
-
-        if outpath:
-            plt.savefig(outpath)
-
-        return projection, ax
-
-    def plot_region(self, hsp_map, region, projection=None, outpath=None, title=None):
-
-        return self.plot_area(
-            hsp_map,
-            region["ra_0"],
-            region["extend"],
-            region["vmax"],
-            projection=projection,
-            outpath=outpath,
-            title=title,
-        )
-
-    def plot_all_regions(self, hsp_map, outbase=None):
-
-        for region in self._regions:
-            if outbase:
-                outpath = f"{outbase}_{region}.png"
-            else:
-                outpath = None
-            self.plot_region(hsp_map, self._regions[region], outpath=outpath)
-
-    @classmethod
-    def hp_pixel_centers(cls, nside, nest=False):
-
-        # Get number of pixels for given nside
-        npix = hp.nside2npix(nside)
-
-        # Get pixel indices
-        pix_indices = np.arange(npix)
-
-        # Get coordinates of pixel centers
-        ra, dec = hp.pix2ang(nside, pix_indices, nest=nest, lonlat=True)
-
-        return ra, dec, npix
-
-    @classmethod
-    def plot_footprint_as_hp(cls, hsp_map, nside, outpath=None, title=None):
-
-        ra, dec, npix = cls.hp_pixel_centers(nside)
-
-        # Create an empty HEALPix map
-        m = np.full(npix, np.nan)
-
-        fig, ax = plt.subplots(figsize=(10, 10))
-
-        # Plot the HEALPix grid
-        hp.mollview(m, title=title, coord="C", notext=True, rot=(180, 0, 0))
-
-        # Define the Galactic Plane: l = [0, 360], b = 0°
-        for l0, ls in zip((-5, 0, 5), (":", "-", ":")):
-            l_values = np.linspace(0, 360, 500)  # 500 points along the plane
-            b_values = np.zeros_like(l_values)  # Galactic latitude is 0 (the plane)
-
-            # Convert (l, b) to (λ, β) - Ecliptic coordinates
-            coords = SkyCoord(
-                l=l_values * u.degree, b=b_values * u.degree, frame="galactic"
-            )
-            ecl_coords = coords.transform_to(
-                "barycentrictrueecliptic"
-            )  # Ecliptic frame
-
-            # Extract Ecliptic longitude (λ) and latitude (β)
-            lambda_ecl = ecl_coords.lon.deg  # Ecliptic longitude
-            beta_ecl = ecl_coords.lat.deg  # Ecliptic latitude
-
-            # Convert to HEALPix projection coordinates (colatitude, longitude)
-            theta = np.radians(90 - beta_ecl)  # HEALPix uses colatitude
-            phi = np.radians(lambda_ecl)  # HEALPix uses longitude
-
-            # Create a healpy Mollweide projection in Ecliptic coordinates
-            hp.projplot(
-                theta, phi, linestyle=ls, color="black", linewidth=1
-            )  # Plot the outline
-
-        # Apply mask
-        mask_values = hsp_map.get_values_pos(ra, dec, valid_mask=True, lonlat=True)
-
-        ok = np.where(mask_values == False)[0]
-        # nok = np.where(mask_values == False)[0]
-
-        hp.projscatter(ra[ok], dec[ok], lonlat=True, color="green", s=1, marker=".")
-        # hp.projscatter(ra[nok], dec[nok], lonlat=True, color="red", s=1, marker=".")
-
-        plt.tight_layout()
-
-        if outpath:
-            plt.savefig(outpath)
-
-        plt.show()
-
-
-def hsp_map_logical_or(maps, verbose=False):
-    """
-    Hsp Map Logical Or.
-
-    Logical AND of HealSparseMaps.
-
-    """
-    if verbose:
-        print("Combine all maps...")
-
-    # Ensure consistency in coverage and data type
-    nside_coverage = maps[0].nside_coverage
-    nside_sparse = maps[0].nside_sparse
-    dtype = maps[0].dtype
-
-    for m in maps:
-        # MKDEBUG TODO: Change nside if possible
-        if m.nside_coverage != nside_coverage:
-            raise ValueError(
-                f"Coverage nside={m.nside_coverage} does not match {nside_coverage}"
-            )
-        if m.dtype != dtype:
-            raise ValueError(f"Data type {m.dtype} does not match {dtype}")
-
-    # Create an empty HealSparse map
-    map_comb = hsp.HealSparseMap.make_empty(nside_coverage, nside_sparse, dtype=dtype)
-    for idx, m in enumerate(maps):
-        map_comb |= m
-
-        if verbose:
-            valid_pixels = map_comb.valid_pixels
-            n_tot = np.sum(valid_pixels)
-            n_true = np.count_nonzero(valid_pixels)
-            n_false = n_tot - n_true
-            print(
-                f"after map {idx}: frac_true={n_true / n_tot:g}, frac_false={n_false / n_tot:g}"
-            )
-
-    return map_comb
